pyproject.toml
[project]
name = "tag-wrangler"
version = "0.1.0"
description = "Streamlit app to manage Obsidian vault tags and frontmatter"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
    "pandas>=2.2.0",
    "plotly>=5.24.0",
    "rapidfuzz>=3.10.0",
]

[project.scripts]
tag-wrangler = "tag_wrangler.app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tag_wrangler"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
